Don't insert (?:) at the beginning/end of a group
This passes the tests in the previous commit, using the new
isPatternNext function to determine if the match is at the end of a
group.

Checking if the match is at the beginning of a group is a little
more naive, since it only looks at the previous character, rather than
ignoring comments and whitespace, but I haven't found a good way to
improve on that.